+++ csharp/src/Ice/OutgoingResponseFrame.cs
@@ -11,66 +11,6 @@ namespace Ice
 {
     using Context = Dictionary<string, string>;
 
-    /// <summary>Represents a response protocol frame received by the application.</summary>
-    // TODO: IncomingResponseFrame should derive from InputStream
-    public sealed class IncomingResponseFrame
-    {
-        // TODO: missing RequestId, Connection, Encoding?
-
-        public InputStream InputStream { get; }
-
-        /// <summary>The Ice1 reply status. Only meaningful for the Ice1 protocol, always set to OK with Ice2.</summary>
-        // TODO: currently it's only OK or UserException as we "throw" other responses.
-        public ReplyStatus ReplyStatus { get; }
-
-        /// <summary>The response context. Always null with Ice1.</summary>
-        public Context? Context { get; }
-
-        /// <summary>The payload of this response frame. The bytes inside the payload should not be written to;
-        /// they are writable because of the <see cref="System.Net.Sockets.Socket"/> methods for sending.</summary>
-        public ArraySegment<byte> Payload
-        {
-            get
-            {
-                if (_payload == null)
-                {
-                    // TODO, it should never be empty, but currently it is when fulfilled by Sent
-                    if (InputStream.IsEmpty)
-                    {
-                        _payload = ArraySegment<byte>.Empty;
-                    }
-                    else
-                    {
-                        // TODO: for other reply status,return the non-encaps payload.
-                        Debug.Assert(ReplyStatus == ReplyStatus.OK || ReplyStatus == ReplyStatus.UserException);
-
-                        // TODO: works only when Payload called first before reading anything. Need a better version!
-                        // TODO: not efficient to create an array here
-                        // TODO: provide Encoding property
-                        _payload = new ArraySegment<byte>(InputStream.ReadEncapsulation(out EncodingVersion _));
-                    }
-                }
-                return _payload.Value;
-            }
-        }
-
-        private ArraySegment<byte>? _payload;
-
-        /// <summary>Take the payload from this response frame. After calling this method, the payload can no longer
-        /// be read.</summary>
-        public ArraySegment<byte> TakePayload()
-        {
-            // TODO: make this method destructive with memory pooling.
-            return Payload;
-        }
-
-        internal IncomingResponseFrame(ReplyStatus replyStatus, InputStream inputStream)
-        {
-            ReplyStatus = replyStatus;
-            InputStream = inputStream;
-        }
-    }
-
     /// <summary>Represents a response protocol frame sent by the application.</summary>
     public sealed class OutgoingResponseFrame : OutputStream
     {
@@ -152,7 +92,7 @@ namespace Ice
         /// <param name="current">The current parameter holds decoded header data and other information about the
         /// request for which this method creates a response.</param>
         /// <returns>A new OutgoingResponseFrame.</returns>
-        public static OutgoingResponseFrame CreateFailure(System.Exception exception, Ice.Current current)
+        public static OutgoingResponseFrame CreateFailure(System.Exception exception, Current current)
         {
             OutgoingResponseFrame responseFrame;
 
@@ -160,7 +100,7 @@ namespace Ice
             {
                 throw exception;
             }
-            catch (Ice.RequestFailedException ex)
+            catch (RequestFailedException ex)
             {
                 if (ex.Id.Name == null || ex.Id.Name.Length == 0)
                 {
@@ -179,15 +119,15 @@ namespace Ice
 
                 ReplyStatus replyStatus = default;
 
-                if (ex is Ice.ObjectNotExistException)
+                if (ex is ObjectNotExistException)
                 {
                     replyStatus = ReplyStatus.ObjectNotExistException;
                 }
-                else if (ex is Ice.FacetNotExistException)
+                else if (ex is FacetNotExistException)
                 {
                     replyStatus = ReplyStatus.FacetNotExistException;
                 }
-                else if (ex is Ice.OperationNotExistException)
+                else if (ex is OperationNotExistException)
                 {
                     replyStatus = ReplyStatus.OperationNotExistException;
                 }
@@ -211,25 +151,25 @@ namespace Ice
                 }
                 responseFrame.WriteString(ex.Operation);
             }
-            catch (Ice.UnknownLocalException ex)
+            catch (UnknownLocalException ex)
             {
                 responseFrame = new OutgoingResponseFrame(current.Adapter.Communicator, current.RequestId,
                     ReplyStatus.UnknownLocalException);
                 responseFrame.WriteString(ex.Unknown);
             }
-            catch (Ice.UnknownUserException ex)
+            catch (UnknownUserException ex)
             {
                 responseFrame = new OutgoingResponseFrame(current.Adapter.Communicator, current.RequestId,
                     ReplyStatus.UnknownUserException);
                 responseFrame.WriteString(ex.Unknown);
             }
-            catch (Ice.UnknownException ex)
+            catch (UnknownException ex)
             {
                responseFrame = new OutgoingResponseFrame(current.Adapter.Communicator, current.RequestId,
                     ReplyStatus.UnknownException);
                 responseFrame.WriteString(ex.Unknown);
             }
-            catch (Ice.UserException ex)
+            catch (UserException ex)
             {
                 responseFrame = StartFailure(current);
                 responseFrame.WriteException(ex);
@@ -259,7 +199,7 @@ namespace Ice
         }
 
         private OutgoingResponseFrame(Communicator communicator, int requestId, ReplyStatus replyStatus)
-            : base(communicator, Ice.Util.CurrentProtocolEncoding)
+            : base(communicator, Util.CurrentProtocolEncoding)
         {
             RequestId = requestId;
             ReplyStatus = replyStatus;
